emit: a sub-int operator result narrows to its declared type
`255.u8 + 1.u8` had two answers. Bound to a local or passed as a u8 argument it
was 0; interpolated, nested in another operator, or converted it was 256. And
`if (a + one) == 0.u8` took the ELSE branch -- a silent wrong turn, not a
printing artifact.

C is the reason. `uint8_t + uint8_t` promotes both operands to `int`, so the
fragment yields an `int` of 256. Wherever the result flowed into a declared
slot the assignment narrowed it back and the answer was right; wherever it was
consumed in place, the promoted value survived. The declaration says
`+: function {:this rhs: type} out type`, so the narrowing is not optional --
the fragment owes it.

40 rows in natives.tbl gain a cast to the declared C type: `+ - * / % & | ^` on
the five widths narrower than `int` -- u8, c8, u16, i8, i16. u32, i32, u64 and
i64 were never affected, which is exactly what C's promotion rules predict and
what measuring confirmed. `<< >> not` were already correct: the bitwise arc cast
them when it added them, and that is the shape the other eight now match.

`&`, `|` and `^` cannot produce an out-of-range value, so their casts change no
result -- they are there because the expression's C TYPE was wrong too, and a
rule with three exceptions is worse than a rule.

examples/narrow_widths.z pins bound == inline across all four affected widths
plus the argument, comparison, nesting and conversion positions. Every operand
in it is a VARIABLE on purpose: two constants fold at compile time, and the
constant evaluator works in i64 by design -- ztypecheck's own comment says a
value carries no target type until it materialises, so a narrower target's
overflow is the materialiser's business. `200.u8 * 2.u8` still folds to 400
rather than 144. That is a separate defect in the materialiser, and whether it
should narrow or refuse is a design question, so it is not answered here.

The compiler's own C moves 14 lines, all of them casts at sites whose values
already fitted -- no behaviour change, which is why the corpus stayed green
before the seed was regenerated. make ci green, seed current.

+++ bootstrap/zc.c
@@ -43703,7 +43703,7 @@ uint32_t z_t1712(z_t78_t z_v17) {
                 uint8_t z_v21 = _m0.data.some;
                 (void)z_v21;
                 if ((z_v21 >= 48) && (z_v21 <= 57)) {
-                    uint32_t z_v22 = ((uint32_t)(z_v21 - 48));
+                    uint32_t z_v22 = ((uint32_t)((uint8_t)(z_v21 - 48)));
                     z_v18 = (z_v18 * 10);
                     z_v18 = (z_v18 + z_v22);
                 }
@@ -48231,7 +48231,7 @@ z_t62_t z_t3075(z_t78_t z_v22868) {
                 (void)z_v22872;
                 if (z_v22872 >= 97) {
                     if (z_v22872 <= 122) {
-                        (void)(z_t62_append_byte(&z_v22869, (z_v22872 - 32)));
+                        (void)(z_t62_append_byte(&z_v22869, ((uint8_t)(z_v22872 - 32))));
                     } else {
                         (void)(z_t62_append_byte(&z_v22869, z_v22872));
                     }
@@ -75662,7 +75662,7 @@ z_t62_t z_t2810(z_t78_t z_v33210) {
             case Z_OPTIONVAL_U8_TAG_SOME: {
                 uint8_t z_v33214 = _m0.data.some;
                 (void)z_v33214;
-                z_v33211 = ((z_v33211 * 10) + ((uint64_t)(z_v33214 - 48)));
+                z_v33211 = ((z_v33211 * 10) + ((uint64_t)((uint8_t)(z_v33214 - 48))));
                 break;
             }
             case Z_OPTIONVAL_U8_TAG_NONE: {
@@ -87577,7 +87577,7 @@ z_t883_t z_t2574(z_t78_t z_v5454) {
                 uint8_t z_v5459 = _m0.data.some;
                 (void)z_v5459;
                 if ((z_v5459 >= 48) && (z_v5459 <= 57)) {
-                    uint8_t z_v5460 = (z_v5459 - 48);
+                    uint8_t z_v5460 = ((uint8_t)(z_v5459 - 48));
                     z_v5455 = ((z_v5455 * 10) + ((uint64_t)z_v5460));
                     z_v5457 = ((bool)Z_BOOL_TAG_TRUE);
                 } else {
@@ -123587,7 +123587,7 @@ bool z_t2344(uint8_t z_v2432) {
         return ((bool)Z_BOOL_TAG_FALSE);
     }
     uint64_t z_v2433 = ((uint64_t)z_v2432);
-    return ((z_t2522[z_v2433] & 0b0001) != 0);
+    return (((uint8_t)(z_t2522[z_v2433] & 0b0001)) != 0);
 }
 
 bool z_t2343(uint8_t z_v2430) {
@@ -123595,7 +123595,7 @@ bool z_t2343(uint8_t z_v2430) {
         return ((bool)Z_BOOL_TAG_FALSE);
     }
     uint64_t z_v2431 = ((uint64_t)z_v2430);
-    return ((z_t2522[z_v2431] & 0b0010) != 0);
+    return (((uint8_t)(z_t2522[z_v2431] & 0b0010)) != 0);
 }
 
 bool z_t2335(uint8_t z_v2193) {
@@ -123603,7 +123603,7 @@ bool z_t2335(uint8_t z_v2193) {
         return ((bool)Z_BOOL_TAG_TRUE);
     }
     uint64_t z_v2194 = ((uint64_t)z_v2193);
-    return ((z_t2522[z_v2194] & 0b0100) != 0);
+    return (((uint8_t)(z_t2522[z_v2194] & 0b0100)) != 0);
 }
 
 bool z_t2336(uint8_t z_v2209) {
@@ -123611,7 +123611,7 @@ bool z_t2336(uint8_t z_v2209) {
         return ((bool)Z_BOOL_TAG_TRUE);
     }
     uint64_t z_v2210 = ((uint64_t)z_v2209);
-    return ((z_t2522[z_v2210] & 0b1000) != 0);
+    return (((uint8_t)(z_t2522[z_v2210] & 0b1000)) != 0);
 }
 
 void z_t2337(z_t62_t* z_v2223, z_t78_t z_v2224) {
@@ -123652,12 +123652,12 @@ bool z_t2338(uint8_t z_v2269) {
 
 uint8_t z_t2339(uint8_t z_v2287) {
     if (z_v2287 <= 57) {
-        return (z_v2287 - 48);
+        return ((uint8_t)(z_v2287 - 48));
     }
     if (z_v2287 <= 70) {
-        return ((z_v2287 - 65) + 10);
+        return ((uint8_t)(((uint8_t)(z_v2287 - 65)) + 10));
     }
-    return ((z_v2287 - 97) + 10);
+    return ((uint8_t)(((uint8_t)(z_v2287 - 97)) + 10));
 }
 
 void z_t2340(uint64_t z_v2292, z_t62_t* z_v2293) {
@@ -124376,17 +124376,17 @@ bool z_t3632(double z_v6979, uint8_t z_v6980) {
 z_t883_t z_t3380(uint8_t z_v6967) {
     if (z_v6967 >= 48) {
         if (z_v6967 <= 57) {
-            return ((z_t883_t){ .tag = Z_OPTIONVAL_U64_TAG_SOME, .data.some = (((uint64_t)(z_v6967 - 48))) });
+            return ((z_t883_t){ .tag = Z_OPTIONVAL_U64_TAG_SOME, .data.some = (((uint64_t)((uint8_t)(z_v6967 - 48)))) });
         }
     }
     if (z_v6967 >= 97) {
         if (z_v6967 <= 102) {
-            return ((z_t883_t){ .tag = Z_OPTIONVAL_U64_TAG_SOME, .data.some = ((((uint64_t)(z_v6967 - 97)) + 10)) });
+            return ((z_t883_t){ .tag = Z_OPTIONVAL_U64_TAG_SOME, .data.some = ((((uint64_t)((uint8_t)(z_v6967 - 97))) + 10)) });
         }
     }
     if (z_v6967 >= 65) {
         if (z_v6967 <= 70) {
-            return ((z_t883_t){ .tag = Z_OPTIONVAL_U64_TAG_SOME, .data.some = ((((uint64_t)(z_v6967 - 65)) + 10)) });
+            return ((z_t883_t){ .tag = Z_OPTIONVAL_U64_TAG_SOME, .data.some = ((((uint64_t)((uint8_t)(z_v6967 - 65))) + 10)) });
         }
     }
     return ((z_t883_t){ .tag = Z_OPTIONVAL_U64_TAG_NONE });
